+++ tests/cli/utils/raw_backup_helpers.py
@@ -5,6 +5,9 @@ import src.utils.common
 from tests.cli.utils.helpers import TEST_PROJECT_NAME
 
 
+SUBPATH_TO_RESTORE = "some-dir"
+
+
 def raw_remote_content_root(
     backup_name: str,
     *,
@@ -17,7 +20,7 @@ def raw_remote_content_root(
         / TEST_PROJECT_NAME
         / backup_name
         / "files"
-        / str(src.utils.common.dir_from_path(str(local_data_dir), deep=deep))
+        / src.utils.common.dir_from_path(str(local_data_dir), deep=deep)
     )
 
 
@@ -25,6 +28,8 @@ def when_having_initial_source_files_for_raw_backup(
     *, doco_config_path: pathlib.Path, local_data_dir: pathlib.Path, local_instance_workdir: pathlib.Path
 ):
     (local_data_dir / "initial_file.txt").write_text("Initial content")
+    (local_data_dir / SUBPATH_TO_RESTORE).mkdir()
+    (local_data_dir / SUBPATH_TO_RESTORE / "some-file").write_text("some file")
 
     shutil.copy(doco_config_path, local_instance_workdir / "doco.config.toml")